+++ src/classes/box.ts
@@ -11,7 +11,6 @@ export class BoxBase extends RectShape {
       fillColor?: string;
       strokeColor?: string;
       padding?: EdgeInsets;
-      margin?: EdgeInsets;
       child: CanvasWidget;
     }
   ) {
@@ -20,37 +19,15 @@ export class BoxBase extends RectShape {
 
   child = this.options.child;
   padding = this.options.padding;
-  margin = this.options.margin;
 
   draw(ctx: CanvasRenderingContext2D, size: Size): void {
     let childSize = { ...size };
-    if (this.margin) {
-      if (Array.isArray(this.margin)) {
-        const [top, left, bottom, right] = this.margin;
-        ctx.translate(left, top);
-        childSize.width -= left + right;
-        childSize.height -= top + bottom;
-      } else {
-        const all = this.margin;
-        ctx.translate(all, all);
-        childSize.width -= all * 2;
-        childSize.height -= all * 2;
-      }
-    }
     super.draw(ctx, childSize);
-    if (this.padding) {
-      if (Array.isArray(this.padding)) {
-        const [top, left, bottom, right] = this.padding;
-        ctx.translate(left, top);
-        childSize.width -= left + right;
-        childSize.height -= top + bottom;
-      } else {
-        const all = this.padding;
-        ctx.translate(all, all);
-        childSize.width -= all * 2;
-        childSize.height -= all * 2;
-      }
-    }
+    edgeInset(this.padding, (top, left, bottom, right) => {
+      ctx.translate(left, top);
+      childSize.width -= left + right;
+      childSize.height -= top + bottom;
+    });
     this.child.draw(ctx, childSize, this.rect);
   }
 
@@ -59,32 +36,21 @@ export class BoxBase extends RectShape {
     selection: CanvasWidget[],
     hovered: CanvasWidget[]
   ) {
-    const rect = this.rect;
-    if (this.margin) {
-      if (Array.isArray(this.margin)) {
-        const [top, left, bottom, right] = this.margin;
-        ctx.translate(-left, -top);
-      } else {
-        const all = this.margin;
-        ctx.translate(-all, -all);
-      }
-    }
-    if (this.padding) {
-      if (Array.isArray(this.padding)) {
-        const [top, left] = this.padding;
-        ctx.translate(-left, -top);
-      } else {
-        const all = this.padding;
-        ctx.translate(-all, -all);
-      }
-    }
+    edgeInset(this.padding, (top, left) => {
+      ctx.translate(-left, -top);
+    });
     if (selection.length > 0 && selection.includes(this)) {
-      drawOutline(ctx, rect, "--canvas-selected-color");
+      drawOutline(ctx, this.rect, "--canvas-selected-color");
     } else if (hovered.length > 0 && hovered.includes(this)) {
-      drawOutline(ctx, rect, "--canvas-hovered-color");
+      drawOutline(ctx, this.rect, "--canvas-hovered-color");
     }
-
-    this.child.drawDecoration(ctx, selection, hovered);
+    let childSize = { width: this.rect.width, height: this.rect.height };
+    edgeInset(this.padding, (top, left, bottom, right) => {
+      ctx.translate(left, top);
+      childSize.width -= left + right;
+      childSize.height -= top + bottom;
+    });
+    this.child.drawDecoration(ctx, selection, hovered, childSize);
   }
 
   selectAt(point: DOMPoint, level: number): CanvasWidget {
@@ -100,17 +66,17 @@ export class BoxBase extends RectShape {
   }
 }
 
-function edgeInset(options: {
-  inset?: EdgeInsets;
-  callback: (top: number, left: number, bottom: number, right: number) => void;
-}) {
-  if (options.inset) {
-    if (Array.isArray(options.inset)) {
-      const [top, left, bottom, right] = options.inset;
-      options.callback(top, left, bottom, right);
+function edgeInset(
+  inset: EdgeInsets | undefined,
+  callback: (top: number, left: number, bottom: number, right: number) => void
+) {
+  if (inset) {
+    if (Array.isArray(inset)) {
+      const [top, left, bottom, right] = inset;
+      callback(top, left, bottom, right);
     } else {
-      const all = options.inset;
-      options.callback(all, all, all, all);
+      const all = inset;
+      callback(all, all, all, all);
     }
   }
 }