+++ src/classes/widget.ts
@@ -1,7 +1,6 @@
-import { drawOutline, Size } from "../utils.js";
+import { drawOutline, Offset, Rect, Size } from "../utils.js";
 
 export abstract class CanvasWidget {
-  abstract rect: DOMRect;
   abstract draw(ctx: CanvasRenderingContext2D, size: Size, parent?: Size): void;
 
   selectAt(point: DOMPoint, level: number): CanvasWidget | null {
@@ -20,12 +19,42 @@ export abstract class CanvasWidget {
   drawDecoration(
     ctx: CanvasRenderingContext2D,
     selection: CanvasWidget[],
-    hovered: CanvasWidget[]
+    hovered: CanvasWidget[],
+    size?: Size
   ) {
+    const rect = size ?? this.rect;
     if (selection.length > 0 && selection.includes(this)) {
-      drawOutline(ctx, this.rect, "--canvas-selected-color");
+      drawOutline(ctx, rect, "--canvas-selected-color");
     } else if (hovered.length > 0 && hovered.includes(this)) {
-      drawOutline(ctx, this.rect, "--canvas-hovered-color");
+      drawOutline(ctx, rect, "--canvas-hovered-color");
     }
   }
+
+  move(delta: Offset) {
+    this.rect.x += delta.x;
+    this.rect.y += delta.y;
+  }
+
+  resize(size: Size) {
+    this.rect.width = size.width;
+    this.rect.height = size.height;
+  }
+
+  get size(): Size {
+    return {
+      width: this.rect.width,
+      height: this.rect.height,
+    };
+  }
+
+  get offset(): Offset {
+    return {
+      x: this.rect.x,
+      y: this.rect.y,
+    };
+  }
+
+  inflate(parent?: Rect): void {}
+
+  abstract get rect(): Rect;
 }