+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Adding padding and margin to box
 - Adding image filter
 - Adding zoomIn, zoomOut, reset
+- Adding padding to box
+- New resize, move and rect getters to widget
 
 ## 0.0.9